Supervisor issues OS shutdown only in Windows.
On Linux, powerdown is done with a script that starts the
Supervisor.

+++ src/main/python/Host/Supervisor/Supervisor.py
@@ -1506,10 +1506,16 @@ class Supervisor(object):
         #Get rid of any backup supervisor first...
         if self.BackupExists:
             self.BackupApp.ShutDown(_METHOD_KILLFIRST,NoKillByName=True,NoWait=True)
-        # Initiate system shutdown
+
+        # If running on Windows initiate system shutdown.
+        # If running on Linux, OS shutdown is managed by a script
+        # that starts the Supervisor.
         if self.powerDownAfterTermination:
-            print("In Supervisor.py calling OS shutdown utility")
-            os.system("shutdown -f -s -t 20")
+            if sys.platform == "win32"
+                os.system("shutdown -f -s -t 20")
+            else:
+                pass
+
         #Now shut applications down in the reverse order of launching...
         #for severity in [_METHOD_STOPFIRST,_METHOD_KILLFIRST,_METHOD_DESTROYFIRST]:
         for severity in [_METHOD_KILLFIRST,_METHOD_DESTROYFIRST]:
